Day 7 part 1
I decided for the nonstreet parts I'm going to group it up by blocks and use perlin noise on the blocks to classify each block. I'm thinking about the detail I want to add such as mixing block and park areas as well as the elevation of the buildings. I'm also thinking about the visuals.